Refactor assignment filtering functions
